refactor metric comparisons
