+++ tox.ini
@@ -90,9 +90,11 @@ commands =
         dadapy/hamming.py \
         dadapy/id_discrete.py \
         dadapy/id_estimation.py \
+        dadapy/information_imbalance.py \
         dadapy/kstar.py \
         dadapy/metric_comparisons.py \
         dadapy/neigh_graph.py \
+        dadapy/neighborhood_overlap.py \
         dadapy/plot.py \
         tests
